opsdb.queue.mjd_plan needs double precision

+++ schema/sdss5db/opsdb/opsdb.sql
@@ -72,7 +72,7 @@ CREATE TABLE opsdb.queue(
     pk SERIAL PRIMARY KEY NOT NULL,
     design_pk INTEGER,
     position SMALLINT,
-    mjd_plan REAL);
+    mjd_plan DOUBLE PRECISION);
 
 -- Foreign keys
